refactor: 파일 구조 변경, 코드 리펙토링
