+++ durihana/types/Schedule.ts
@@ -1,0 +1,12 @@
+export type Schedule = {
+  id: number;
+  title: string;
+  date: Date;
+  time: string;
+  type: 'reservation' | 'finance';
+  partnerName?: string;
+};
+
+export type ScheduleProps = {
+  userId: number;
+};
